AAVS-469: Finalise implementation after testing with real data.

+++ docroot/modules/custom/alshaya_product_options/src/SwatchesHelper.php
@@ -94,7 +94,9 @@ class SwatchesHelper {
       if ($term->get('field_attribute_swatch_image')->first()) {
         $file_value = $term->get('field_attribute_swatch_image')->first()->getValue();
         $file = $this->fileStorage->load($file_value['target_id']);
-        $file->delete();
+        if ($file) {
+          $file->delete();
+        }
       }
 
       // Reset all current values.
